# PERF: high memory in MI