--- a/pandas/src/hashtable_class_helper.pxi.in
+++ b/pandas/src/hashtable_class_helper.pxi.in
@@ -262,13 +262,6 @@ cdef class {{name}}HashTable(HashTable):
         else:
             raise KeyError(val)
 
-    def get_iter_test(self, {{dtype}}_t key, Py_ssize_t iterations):
-        cdef Py_ssize_t i, val=0
-        for i in range(iterations):
-            k = kh_get_{{dtype}}(self.table, val)
-            if k != self.table.n_buckets:
-                val = self.table.vals[k]
-
     cpdef set_item(self, {{dtype}}_t key, Py_ssize_t val):
         cdef:
             khiter_t k
@@ -500,18 +493,6 @@ cdef class StringHashTable(HashTable):
             return self.table.vals[k]
         else:
             raise KeyError(val)
-
-    def get_iter_test(self, object key, Py_ssize_t iterations):
-        cdef:
-            Py_ssize_t i, val
-            char *v
-
-        v = util.get_c_string(key)
-
-        for i in range(iterations):
-            k = kh_get_str(self.table, v)
-            if k != self.table.n_buckets:
-                val = self.table.vals[k]
 
     cpdef set_item(self, object key, Py_ssize_t val):
         cdef:
@@ -755,15 +736,6 @@ cdef class PyObjectHashTable(HashTable):
         else:
             raise KeyError(val)
 
-    def get_iter_test(self, object key, Py_ssize_t iterations):
-        cdef Py_ssize_t i, val
-        if key != key or key is None:
-            key = na_sentinel
-        for i in range(iterations):
-            k = kh_get_pymap(self.table, <PyObject*>key)
-            if k != self.table.n_buckets:
-                val = self.table.vals[k]
-
     cpdef set_item(self, object key, Py_ssize_t val):
         cdef:
             khiter_t k
@@ -874,3 +846,127 @@ cdef class PyObjectHashTable(HashTable):
                 count += 1
 
         return np.asarray(labels)
+
+
+cdef class MultiIndexHashTable(HashTable):
+
+    def __init__(self, size_hint=1):
+        self.table = kh_init_uint64()
+        self.mi = None
+        kh_resize_uint64(self.table, size_hint)
+
+    def __dealloc__(self):
+        if self.table is not NULL:
+            kh_destroy_uint64(self.table)
+            self.table = NULL
+
+    def __len__(self):
+        return self.table.size
+
+    def sizeof(self, deep=False):
+        """ return the size of my table in bytes """
+        return self.table.n_buckets * (sizeof(uint64_t) + # keys
+                                       sizeof(size_t) + # vals
+                                       sizeof(uint32_t)) # flags
+
+    def _check_for_collisions(self, int64_t[:] locs, object mi):
+        # validate that the locs map to the actual values
+        # provided in the mi
+        # we can only check if we *don't* have any missing values
+        # :<
+        cdef:
+            ndarray[int64_t] alocs
+
+        alocs = np.asarray(locs)
+        if (alocs != -1).all():
+
+            result = self.mi.take(locs)
+            if isinstance(mi, tuple):
+                from pandas import Index
+                mi = Index([mi])
+            if not result.equals(mi):
+                raise AssertionError(
+                    "hash collision\nlocs:\n{}\n"
+                    "result:\n{}\nmi:\n{}".format(alocs, result, mi))
+
+    def __contains__(self, object key):
+        try:
+            self.get_item(key)
+            return True
+        except (KeyError, ValueError, TypeError):
+            return False
+
+    cpdef get_item(self, object key):
+        cdef:
+            khiter_t k
+            uint64_t value
+            int64_t[:] locs
+            Py_ssize_t loc
+
+        value = self.mi._hashed_indexing_key(key)
+        k = kh_get_uint64(self.table, value)
+        if k != self.table.n_buckets:
+            loc = self.table.vals[k]
+            locs = np.array([loc], dtype=np.int64)
+            self._check_for_collisions(locs, key)
+            return loc
+        else:
+            raise KeyError(key)
+
+    cpdef set_item(self, object key, Py_ssize_t val):
+        raise NotImplementedError
+
+    @cython.boundscheck(False)
+    def map_locations(self, object mi):
+        cdef:
+            Py_ssize_t i, n
+            ndarray[uint64_t] values
+            uint64_t val
+            int ret = 0
+            khiter_t k
+
+        self.mi = mi
+        n = len(mi)
+        values = mi._hashed_values
+
+        with nogil:
+            for i in range(n):
+                val = values[i]
+                k = kh_put_uint64(self.table, val, &ret)
+                self.table.vals[k] = i
+
+    @cython.boundscheck(False)
+    def lookup(self, object mi):
+        # look up with a target mi
+        cdef:
+            Py_ssize_t i, n
+            ndarray[uint64_t] values
+            int ret = 0
+            uint64_t val
+            khiter_t k
+            int64_t[:] locs
+
+        n = len(mi)
+        values = mi._hashed_values
+
+        locs = np.empty(n, dtype=np.int64)
+
+        with nogil:
+            for i in range(n):
+                val = values[i]
+                k = kh_get_uint64(self.table, val)
+                if k != self.table.n_buckets:
+                    locs[i] = self.table.vals[k]
+                else:
+                    locs[i] = -1
+
+        self._check_for_collisions(locs, mi)
+        return np.asarray(locs)
+
+    def unique(self, object mi):
+        raise NotImplementedError
+
+    def get_labels(self, object mi, ObjectVector uniques,
+                   Py_ssize_t count_prior, int64_t na_sentinel,
+                   bint check_null=True):
+        raise NotImplementedError