… doc/source/whatsnew/v0.20.0.txt
… pandas/tests/frame/test_mutate_columns.py
… pandas/tests/frame/test_repr_info.py
… pandas/tests/indexes/test_multi.py
… pandas/tests/indexing/test_multiindex.py
… pandas/tests/test_multilevel.py
… pandas/tests/tools/test_hashing.py
… pandas/tests/tools/test_join.py
… pandas/types/cast.py