@jreback 2017-01-31
ok, updated. had an insiduous condition in ``MultiIndex.equals`` with NaT only tuples. This was a latent bug.