@jreback 2017-02-02
ok this is all set. any more comments

@wesm @jorisvandenbossche