@jreback 2017-02-09
@wesm @jorisvandenbossche any comments?