@wesm 2017-02-10
Will take a closer look at this today if possible