@jreback 2017-02-10
@chris-b1 suggestion. not bad.
```
    before     after       ratio
  [dab01796] [403a0864]
-  256.41ms   200.36ms      0.78  indexing.MultiIndexing.time_multiindex_string_get_loc
-  432.21ms   246.25ms      0.57  reindex.Reindexing.time_reindex_multiindex

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```