@jreback 2017-02-10
and vs. master
```
    before     after       ratio
  [e8840725] [403a0864]
-  837.68ms   259.01ms      0.31  reindex.Reindexing.time_reindex_multiindex
-  727.12ms   201.54ms      0.28  indexing.MultiIndexing.time_multiindex_string_get_loc
-  718.79ms   187.85ms      0.26  indexing.MultiIndexing.time_multiindex_get_indexer

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```