@jreback 2017-02-11
@jorisvandenbossche fixed up. ``__contains__`` was tested much, so behavior wasn't nailed down.