@jorisvandenbossche 2017-02-11
They already compare equal, also on your PR branch. So nothing to change
(only the test)

(github is doing annoying, and I cannot find this thread in the web
interface, so trying via email)