@jreback 2017-02-11
We have this in 0.19.2 / master
seems odd
```
In [1]:         midx = MultiIndex.from_product([['A', 'B'], [1, 2]])
   ...:         df = DataFrame(np.random.randn(4, 4), columns=midx)
   ...: 

In [2]: 'A' in df
Out[2]: True

In [3]: ('A', ) in df
Out[3]: True

In [4]: del df['A']

In [5]: 'A' in df
Out[5]: True

In [6]: ('A', ) in df
Out[6]: False
```