@jorisvandenbossche 2017-02-12
In [5] seems certainly like a bug. Possibly because it's looking in the full level and not the actually used one?