@jreback 2017-02-15
all changes pushed. WIll merge on pass tomorrow sometime unless any more comments.