@jreback 2017-03-02
neat article about hash collisions: https://medium.com/@robertgrosse/generating-64-bit-hash-collisions-to-dos-python-5b21404a5306#.wi6ivpqtv